Add UI components for lecturer and student dashboards, quizzes, and messaging
- Created CSS styles for login page.
- Added FXML files for admin dashboard, sidebar, top bar, lecturer dashboard, quiz creation, quiz editing, quiz listing, and student dashboard.
- Implemented messaging interface with user list and chat functionality.
- Developed quiz attempt and result pages for students.
- Enhanced layout and styling for improved user experience across various components.

+++ desktop/E-Discuss/src/main/java/com/edudiscuss/models/ActiveQuizResponse.java
@@ -5,10 +5,10 @@ public class ActiveQuizResponse {
     private boolean active;
     private int quiz_id;
     private String deadline;
-    private int remaining_seconds;
+    private double remaining_seconds;
 
     public boolean isActive() { return active; }
     public int getQuizId() { return quiz_id; }
     public String getDeadline() { return deadline; }
-    public int getRemainingSeconds() { return remaining_seconds; }
+    public double getRemainingSeconds() { return remaining_seconds; }
 }
